requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
chromadb>=0.4.0
openai>=1.0.0

# Document processing
pypdf>=3.17.0
python-docx>=1.0.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Text processing
tiktoken>=0.5.0
